% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteStudy.R
\name{WriteStudy}
\alias{WriteStudy}
\title{Write a BBrowser study}
\usage{
WriteStudy(
  expr.data,
  metadata,
  dimred.data,
  study.path,
  clustering.name,
  author,
  unique.limit
)
}
\arguments{
\item{expr.data}{list of counts, norms, and feature_type}

\item{metadata}{a data.frame of metadata}

\item{dimred.data}{a named list of dimensionality reductions}

\item{study.path}{path to study directory}

\item{clustering.name}{name of metadata that indicates clustering result}

\item{author}{email of the creator}
}
\description{
Write a BBrowser study
}
