% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportSeuratObject.R
\name{ExportSeuratObject}
\alias{ExportSeuratObject}
\title{Export a Seurat object in BCS format}
\usage{
ExportSeuratObject(object, bcs.path)
}
\arguments{
\item{object}{a Seurat object}

\item{bcs.path}{path to BCS file}
}
\description{
Export a Seurat object in BCS format
}
